new/requirements.txt
# Enhanced SAT Vocabulary RAG System - Production Dependencies

# Core Framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-multipart>=0.0.6,<0.1.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0

# Vector Database & Embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2

# HTTP & API
requests>=2.31.0,<3.0.0
httpx>=0.25.0
aiohttp>=3.9.0

# Template & Web
jinja2>=3.1.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database (sqlite3 is built into Python)
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
pyyaml>=6.0.0,<7.0.0
python-dateutil>=2.8.0
uuid
# pathlib is built into Python 3.4+

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pylint>=2.17.0

# Monitoring & Logging
structlog>=23.0.0
prometheus-client>=0.17.0

# Security
cryptography>=41.0.0